feature (migrate token registration to first page)
